refactor: move `penumbra-crypto` staking-related types to stake crate
Unfortunately to do this without introducing a cyclic dependency,
the `Penalty` struct usage had to be removed from `penumbra-chain`.
