+++ crates/bin/pcli/src/command/query/validator.rs
@@ -3,9 +3,11 @@ use std::{fs::File, io::Write};
 use anyhow::{Context, Result};
 use comfy_table::{presets, Table};
 use futures::TryStreamExt;
-use penumbra_crypto::stake::IdentityKey;
 use penumbra_proto::client::v1alpha1::ValidatorInfoRequest;
-use penumbra_stake::validator::{self, ValidatorToml};
+use penumbra_stake::{
+    validator::{self, ValidatorToml},
+    IdentityKey,
+};
 
 use crate::App;
