+++ crates/bin/pcli/src/command/view/staked.rs
@@ -3,11 +3,11 @@ use std::collections::BTreeMap;
 use anyhow::Result;
 use comfy_table::{presets, Table};
 use futures::TryStreamExt;
-use penumbra_crypto::{stake::DelegationToken, FullViewingKey, Value, STAKING_TOKEN_ASSET_ID};
+use penumbra_crypto::{FullViewingKey, Value, STAKING_TOKEN_ASSET_ID};
 use penumbra_proto::client::v1alpha1::{
     oblivious_query_service_client::ObliviousQueryServiceClient, ValidatorInfoRequest,
 };
-use penumbra_stake::validator;
+use penumbra_stake::{validator, DelegationToken};
 use penumbra_view::ViewClient;
 use tonic::transport::Channel;
