+++ crates/bin/pd/src/testnet/generate.rs
@@ -5,12 +5,10 @@ use crate::testnet::{generate_tm_config, parse_tm_address, write_configs, Valida
 use anyhow::{Context, Result};
 use penumbra_chain::genesis;
 use penumbra_chain::{genesis::Allocation, params::ChainParameters};
-use penumbra_crypto::{
-    keys::SpendKey,
-    stake::{DelegationToken, IdentityKey},
-    Address, GovernanceKey,
+use penumbra_crypto::{keys::SpendKey, Address, GovernanceKey};
+use penumbra_stake::{
+    validator::Validator, DelegationToken, FundingStream, FundingStreams, IdentityKey,
 };
-use penumbra_stake::{validator::Validator, FundingStream, FundingStreams};
 use serde::{de, Deserialize};
 use std::{
     fmt,
