+++ crates/core/component/chain/src/params.rs
@@ -6,7 +6,7 @@ use std::{
 };
 
 use anyhow::Context;
-use penumbra_crypto::{stake::Penalty, Amount};
+use penumbra_crypto::Amount;
 use penumbra_proto::client::v1alpha1 as pb_client;
 use penumbra_proto::core::chain::v1alpha1 as pb_chain;
 
@@ -31,9 +31,9 @@ pub struct ChainParameters {
     /// The base reward rate, expressed in basis points of basis points
     pub base_reward_rate: u64,
     /// The penalty for slashing due to misbehavior, expressed in basis points squared (10^-8)
-    pub slashing_penalty_misbehavior: Penalty,
+    pub slashing_penalty_misbehavior: u64,
     /// The penalty for slashing due to downtime, expressed in basis points squared (10^-8)
-    pub slashing_penalty_downtime: Penalty,
+    pub slashing_penalty_downtime: u64,
     /// The number of blocks in the window to check for downtime.
     pub signed_blocks_window_len: u64,
     /// The maximum number of blocks in the window each validator can miss signing without slashing.
@@ -79,8 +79,8 @@ impl TryFrom<pb_chain::ChainParameters> for ChainParameters {
             epoch_duration: msg.epoch_duration,
             unbonding_epochs: msg.unbonding_epochs,
             active_validator_limit: msg.active_validator_limit,
-            slashing_penalty_downtime: Penalty(msg.slashing_penalty_downtime),
-            slashing_penalty_misbehavior: Penalty(msg.slashing_penalty_misbehavior),
+            slashing_penalty_downtime: msg.slashing_penalty_downtime,
+            slashing_penalty_misbehavior: msg.slashing_penalty_misbehavior,
             base_reward_rate: msg.base_reward_rate,
             missed_blocks_maximum: msg.missed_blocks_maximum,
             signed_blocks_window_len: msg.signed_blocks_window_len,
@@ -140,8 +140,8 @@ impl From<ChainParameters> for pb_chain::ChainParameters {
             active_validator_limit: params.active_validator_limit,
             signed_blocks_window_len: params.signed_blocks_window_len,
             missed_blocks_maximum: params.missed_blocks_maximum,
-            slashing_penalty_downtime: params.slashing_penalty_downtime.0,
-            slashing_penalty_misbehavior: params.slashing_penalty_misbehavior.0,
+            slashing_penalty_downtime: params.slashing_penalty_downtime,
+            slashing_penalty_misbehavior: params.slashing_penalty_misbehavior,
             base_reward_rate: params.base_reward_rate,
             ibc_enabled: params.ibc_enabled,
             inbound_ics20_transfers_enabled: params.inbound_ics20_transfers_enabled,
@@ -169,9 +169,9 @@ impl Default for ChainParameters {
             signed_blocks_window_len: 10000,
             missed_blocks_maximum: 9500,
             // 1000 basis points = 10%
-            slashing_penalty_misbehavior: Penalty(1000_0000),
+            slashing_penalty_misbehavior: 1000_0000,
             // 1 basis point = 0.01%
-            slashing_penalty_downtime: Penalty(1_0000),
+            slashing_penalty_downtime: 1_0000,
             // 3bps -> 11% return over 365 epochs
             base_reward_rate: 3_0000,
             ibc_enabled: true,
