+++ crates/core/component/stake/Cargo.toml
@@ -35,6 +35,7 @@ penumbra-chain = { path = "../chain", default-features = false }
 
 # Penumbra dependencies
 decaf377 = {version = "0.4", features = ["r1cs"] }
+decaf377-rdsa = { version = "0.6" }
 ark-r1cs-std = {version = "0.4", default-features = false }
 ark-relations = "0.4"
 ark-ff = { version = "0.4", default_features = false }
