--- crates/core/crypto/src/stake.rs
@@ -1,9 +1,0 @@
-mod delegation_token;
-mod identity_key;
-mod penalty;
-mod unbonding_token;
-
-pub use delegation_token::DelegationToken;
-pub use identity_key::IdentityKey;
-pub use penalty::{Penalty, PenaltyVar};
-pub use unbonding_token::UnbondingToken;
